\name{rescale}
\alias{rescale}
\title{Rescaling as Z-scores and Centering 
}
\description{
}
\usage{
rescale(x, fun = "center")
}
\arguments{
  \item{x}{ A vector to be computed
}
  \item{fun}{A function to be applied: 'z-score', the defaut is 'center'. 
}
}
\details{
}
\value{An object the same class as \code{x}.
}
\references{
}
\author{Daniel Marcelino
}
\note{
}


\seealso{
}
\examples{

}
\keyword{ center mean }
\keyword{ z-scores }

